requirements.txt
numpy
Pillow
PyOpenGL
PyGLM
glfw